pyproject.toml
[tool.poetry]
name = "work_abroad_api"
version = "0.1.0"
description = ""
authors = ["Adam Janas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"
django-phone-field = "^1.8.1"
django-countries = "^6.1.2"
django-rest-auth = "^0.9.5"
django-cleanup = "^5.0.0"
django-filter = "^2.3.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
isort = "^5.0.7"
flake8 = "^3.8.3"
bandit = "^1.6.2"
autoflake = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
